change data type to Gistit
