+++ gistit-ipc/Cargo.toml
@@ -15,6 +15,7 @@ bincode = "1.3.3"
 serde = { version = "1.0.136", features = ["derive"] }
 log = "0.4.14"
 tokio = { version = "1.16.1", default-features = false, features = ["net", "rt", "macros"] }
+gistit-reference = { version = "0.1.0", path = "../gistit-reference" }
 
 [dev-dependencies]
 assert_fs = "1.0.7"
