+++ gistit-ipc/src/lib.rs
@@ -7,6 +7,7 @@
 //         \/        \/
 //
 #![warn(clippy::all, clippy::pedantic, clippy::nursery, clippy::cargo)]
+#![allow(clippy::module_name_repetitions)]
 #![cfg_attr(
     test,
     allow(
@@ -31,8 +32,13 @@ use tokio::net::UnixDatagram;
 use bincode::{deserialize, serialize};
 use serde::{Deserialize, Serialize};
 
-const NAMED_SOCKET_0: &str = "gistit-0";
-const NAMED_SOCKET_1: &str = "gistit-1";
+use gistit_reference::{Gistit, NAMED_SOCKET_0, NAMED_SOCKET_1};
+
+pub mod error;
+pub use bincode;
+
+pub type Result<T> = std::result::Result<T, error::Error>;
+
 const READBUF_SIZE: usize = 60_000; // A bit bigger than 50kb because encoding
 const CONNECT_TIMEOUT_SECS: u64 = 3;
 
@@ -200,7 +206,7 @@ impl Bridge<Client> {
 pub enum Instruction {
     Provide {
         hash: String,
-        data: Vec<u8>,
+        data: Gistit,
     },
     Get {
         hash: String,
@@ -219,53 +225,13 @@ pub enum Instruction {
 #[derive(Serialize, Deserialize, Debug, PartialEq, Eq)]
 pub enum ServerResponse {
     Status {
+        peer_id: String,
         peer_count: usize,
         pending_connections: u32,
         listeners: Vec<String>,
+        hosting: usize,
     },
-    PeerId(String),
     File(Vec<u8>),
-}
-
-pub type Result<T> = std::result::Result<T, Error>;
-
-#[derive(Debug)]
-pub struct Error {
-    pub kind: ErrorKind,
-}
-
-#[derive(Debug)]
-pub enum ErrorKind {
-    IO(std::io::Error),
-    Serialization(bincode::Error),
-}
-
-impl From<std::io::Error> for Error {
-    fn from(err: std::io::Error) -> Self {
-        Self {
-            kind: ErrorKind::IO(err),
-        }
-    }
-}
-
-impl From<bincode::Error> for Error {
-    fn from(err: bincode::Error) -> Self {
-        Self {
-            kind: ErrorKind::Serialization(err),
-        }
-    }
-}
-
-impl std::error::Error for Error {
-    fn description(&self) -> &str {
-        "gistit ipc error"
-    }
-}
-
-impl std::fmt::Display for Error {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "gistit ipc error")
-    }
 }
 
 #[cfg(test)]
